+++ src/App.jsx
@@ -1,17 +1,14 @@
 import './App.css'
+import TodoForm from './TodoForm'
+import TodoList from './TodoList'
 
 function App() {
-  const todos = [
-    {id: 1, title: "review resources"},
-    {id: 2, title: "take notes"},
-    {id: 3, title: "code out app"},
-]
+  
   return (
 <div>
-  <h1>My Todos</h1>
-    <ul>
-      {todos.map(todo => <li key={todo.id}>{todo.title}</li>)}
-    </ul>
+  <h1>My To-do</h1>
+  <TodoForm/>
+  <TodoList/>
 </div>
   )
 }
